Logo Changes, Navbar Fixed & Added onto Route FIle to appear on all Routes. Navbar Data Object removed from Home & added into Navbar Component (Made it Static). Added Style for App css for router to stay at top
