+++ client/src/components/Navbar/Navbar.css
@@ -7,6 +7,7 @@
   justify-content: space-between;
   align-items: center;
   padding: 0 2rem;
+  margin-bottom: 1.5rem;
   font-size: 1.2rem;
 }
